refactor: reorganize downloadPtau import paths for better module structure
